Use a single `errors` data structure throughout a formatToString() call
