+++ icu4c/source/i18n/sources.txt
@@ -96,6 +96,7 @@ measure.cpp
 msgfmt.cpp
 messageformat2.cpp
 messageformat2_checker.cpp
+messageformat2_context.cpp
 messageformat2_data_model.cpp
 messageformat2_formatted_value.cpp
 messageformat2_function_registry.cpp
